# agent-relay/public/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentRelay HTTP API",
    "version": "0.1.0",
    "description": "AgentRelay 测试台后端 REST API 规范。默认基址 `http://localhost:18787`。\n\n- 交互参考：[/api-docs](/api-docs)\n- 说明文档：[/docs](/docs)\n- 机器可读规范：[/api-spec.json](/api-spec.json)\n- 能力探测：`GET /api/config` → `capabilities`",
    "contact": {
      "name": "AgentRelay"
    }
  },
  "servers": [
    {
      "url": "http://localhost:18787",
      "description": "本地测试台（npm run serve）"
    }
  ],
...
  "paths": {
...
    "/api/storage/cleanup/preview": {
      "post": {
        "tags": ["安全与审计"],
        "summary": "清理预览（dry-run，不删除）",
        "operationId": "previewStorageCleanup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scope": { "type": "string", "enum": ["safe", "all"] },
                  "include": { "type": "array", "items": { "type": "string" } },
                  "olderThanDays": { "type": ["number", "null"] },
                  "maxRisk": { "type": "string", "enum": ["low", "medium", "high"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "候选清理动作与预计释放空间",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          }
        }
      }
    },
    "/api/storage/cleanup/apply": {
      "post": {
        "tags": ["安全与审计"],
        "summary": "执行已预览的低风险清理",
        "operationId": "applyStorageCleanup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["cleanupRunId", "confirm"],
                "properties": {
                  "cleanupRunId": { "type": "string" },
                  "confirm": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "清理结果", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "description": "预览不存在" },
          "409": { "description": "清理锁冲突" },
          "410": { "description": "预览已过期" }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": ["对话与智能体"],
        "summary": "单次模型对话",
        "operationId": "chat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "模型回复", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "502": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": ["对话与智能体"],
        "summary": "单次对话 SSE 流式（token + done）",
        "operationId": "chatStream",
        "description": "走 ModelRouter 直连以支持 `streamTokens`；校验失败返回 JSON。事件：run_start、token、done、error。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatStreamRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 事件流",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string", "description": "event: run_start|token|done|error" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/agent/stream": {
      "post": {
        "tags": ["对话与智能体"],
        "summary": "自主 Agent 循环（SSE 思考过程 + 工具步骤）",
        "operationId": "runAgentStream",
        "description": "校验失败时返回 JSON（400/404）；成功时 `text/event-stream`。事件：run_start、model_turn（每轮模型决策）、step、activity_event（公开执行过程）、token（streamTokens=true）、done、error。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AgentRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 事件流",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string", "description": "event: run_start|model_turn|step|activity_event|token|done|error" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/agent/resume": {
      "post": {
        "tags": ["对话与智能体"],
        "summary": "续跑预算耗尽的 Agent Run（RunStateStore）",
        "description": "从 `run_states` 恢复 `pendingSteps`（主要为 PlanWorkflow 的 locate/context_pack），可传更高 `budget` 继续只读预扫描。",
        "operationId": "resumeAgent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AgentResumeRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "续跑结果（含 resumed=true）", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
    "/api/agent": {
      "post": {
        "tags": ["对话与智能体"],
        "summary": "自主 Agent 循环（ReAct JSON）",
        "description": "统一 Agent 入口。模型首轮前的确定性工作流由 WorkflowExecutor 统一调度：PlanWorkflow 负责只读预扫描；debugWorkflow 会先执行 locate_relevant_files + context_pack 只读诊断定位，再由 DebugAnalysisWorkflow 注入 analysis phase（errorSummary/suspectedFiles/rootCauseHypotheses/minimalFixPlan/verificationPlan/riskAndRollback）以约束报错分析，并在 executionMeta.workflowDebugAnalyses 返回可审计诊断记录；refactorWorkflow 会先执行 project_scan + locate_relevant_files + context_pack 只读预扫描，再由 RefactorPlanWorkflow 注入 plan phase（scopeSummary/affectedModules/stagedChanges/perStageVerification/riskAndRollback）强制分阶段重构计划，并在 executionMeta.workflowRefactorPlans 返回可审计记录；复杂副作用任务（未显式进入计划模式）会由 ImplicitPlanWorkflow 注入内部轻量计划（goalSummary/internalSteps/successCriteria/stopConditions，userVisiblePlanMode=false），记录到 executionMeta.workflowInternalPlans；运行结束会在 executionMeta.workflowTaskState 返回 idle/planning/waiting_confirmation/executing/verifying/completed/failed/cancelled；editWorkflow/generateFileWorkflow 会先执行 locate_relevant_files + context_pack 只读预定位，再由 EditProposalWorkflow 注入 proposal phase（targetFiles/changeSummary/permissionCheck/diffPlan/verificationPlan）以约束写入前方案，并在 executionMeta.workflowProposals 返回可审计 proposal 记录和 PermissionGuard 写入前预检结果；成功执行 write_file/apply_patch 后会在 executionMeta.workflowDiffs 汇总 path/changeId/hash/diff 审计摘要，并由 EditExecutionWorkflow 注入 execution phase 上下文，要求下一轮基于真实 diff 做最小验证或最终总结；若写工具返回目标路径，EditAutoVerificationWorkflow 会规划只读 read_file，由 AgentLoop 通过既有权限、预算与工具链路自动读回刚写入文件；写入后的 read_file/diff_file/shell_run 等验证工具结果会由 EditVerificationWorkflow 注入 verification phase，并在 executionMeta.workflowVerifications 返回验证工具、状态、错误与输出预览；WorkflowStateCenter 会从计划、写入、验证、修正事件派生 executionMeta.workflowState；WorkflowWriteGate 基于该状态阻止缺少计划/只读上下文的首次写入、阻止未验证前的连续写入，并要求 refactorWorkflow 先有 staged plan；若验证失败，WorkflowCorrectionWorkflow 会按路径统计修正轮次，注入 correction/termination phase，并在 executionMeta.workflowCorrections 返回 attempt/limitReached 等审计记录（edit/generate-file/debug/refactor 共用，默认达到 2 次失败即终止）；runWorkflow/verifyWorkflow 会尝试执行白名单安全命令（node --version、npm --version、npm run typecheck、npm test、npm run build）；无安全命令、无 shell 权限或预算不足时静态降级并说明。",
        "operationId": "runAgent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AgentRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "最终答案与工具步骤", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "502": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
...
    "/api/plan": {
      "post": {
        "tags": ["计划与任务"],
        "summary": "生成机器计划草案预览（Markdown + PublicPlanJson，不可直接执行）",
        "description": "调用 Planner 生成可持久化的 InternalTaskPlan，并返回 previewMarkdown / publicPlanJson（executable=false）。本接口要求模型输出可解析的计划 JSON；给用户看的 Markdown 计划分析报告请改用 `/api/agent` 且传入 `mode=plan`。真实执行请走 `/api/plans/:planId/approve` → `/api/plans/:planId/execute`。",
        "operationId": "createPlan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlanRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "计划预览", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlanResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
    "/api/plans/draft": {
      "post": {
        "tags": ["计划与任务"],
        "summary": "起草 InternalTaskPlan 并返回展示预览",
        "operationId": "draftPlan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlanRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "planId + previewMarkdown + publicPlanJson", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlanDraftResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
    "/api/plans/analyze": {
      "post": {
        "tags": ["计划与任务"],
        "summary": "生成用户可见计划报告（UserVisiblePlan）",
        "description": "通过 PlanReportWorkflow 走 /api/agent mode=plan 的只读链路，输出并持久化 UserVisiblePlan Markdown；该文档不可直接执行，HTTP handler 仅负责参数校验与模型选择。",
        "operationId": "analyzePlan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlanAnalyzeRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "UserVisiblePlan + executionMeta", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlanAnalyzeResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "502": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" }
          }
        }
      },
...
    },
    "schemas": {
...
      "ChatResponse": {
        "type": "object",
        "properties": {
          "runId": { "type": "string", "format": "uuid" },
          "content": { "type": "string" },
          "routerDecision": { "type": "object", "description": "规则路由摘要（taskType/strategy/模型 id）" },
          "executionStrategy": {
            "type": "string",
            "enum": ["single_model", "strong_model_direct", "local_draft_remote_review"]
          },
          "collaborationRunId": { "type": "string", "description": "多模型协作 run id（如有）" },
          "fallbackCount": { "type": "integer", "description": "V2：本请求触发的模型升级 fallback 次数（0 时省略）" },
          "fallbackLogIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "description": "V2：fallback_logs 记录 id 列表（无 fallback 时省略）"
          },
          "clientName": { "type": "string" },
          "modelName": { "type": "string" },
          "location": { "type": "string" },
          "latencyMs": { "type": "number" },
          "usage": { "type": "object", "additionalProperties": true }
        }
      },
...
    }
  }
}